+++ hois_back/src/main/java/ee/hitsa/ois/web/SubjectController.java
@@ -1,11 +1,29 @@
 package ee.hitsa.ois.web;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import javax.validation.Valid;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
 import ee.hitsa.ois.domain.subject.Subject;
-import ee.hitsa.ois.repository.CurriculumVersionRepository;
 import ee.hitsa.ois.service.AutocompleteService;
 import ee.hitsa.ois.service.SubjectService;
 import ee.hitsa.ois.service.security.HoisUserDetails;
-import ee.hitsa.ois.util.UserUtil;
+import ee.hitsa.ois.util.SubjectUserRights;
+import ee.hitsa.ois.util.SubjectUserRightsValidator;
 import ee.hitsa.ois.util.WithEntity;
 import ee.hitsa.ois.util.WithVersionedEntity;
 import ee.hitsa.ois.web.commandobject.SubjectForm;
@@ -16,16 +34,6 @@ import ee.hitsa.ois.web.dto.AutocompleteResult;
 import ee.hitsa.ois.web.dto.SubjectDto;
 import ee.hitsa.ois.web.dto.SubjectSearchDto;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 @RestController
 @RequestMapping("/subject")
 public class SubjectController {
@@ -34,43 +42,53 @@ public class SubjectController {
     private AutocompleteService autocompleteService;
     @Autowired
     private SubjectService subjectService;
-    @Autowired
-    private CurriculumVersionRepository curriculumVersionRepository;
 
     @PostMapping
     public SubjectDto create(HoisUserDetails user, @Valid @RequestBody SubjectForm newSubject) {
+        SubjectUserRightsValidator.assertCanCreate(user);
         return get(user, subjectService.create(user, newSubject));
+    }
+    
+    @GetMapping("/getPermissions")
+    public Map<String, ?> canCreate(HoisUserDetails user) {
+        SubjectUserRightsValidator.assertCanSearch(user);
+        Map<String, Boolean> response = new HashMap<>();
+        response.put("canCreate", Boolean.valueOf(SubjectUserRights.hasPermissionToEdit(user)));
+        response.put("canViewAll", Boolean.valueOf(SubjectUserRights.canViewAllSubjects(user)));
+        return response;        
     }
 
     @PutMapping("/{id:\\d+}")
-    public SubjectDto save(HoisUserDetails user, 
-            @WithVersionedEntity(value = "id", versionRequestBody = true) Subject subject, 
+    public SubjectDto save(HoisUserDetails user, @WithVersionedEntity(versionRequestBody = true) Subject subject,
             @Valid @RequestBody SubjectForm newSubject) {
-        UserUtil.assertSameSchool(user, subject.getSchool());
+        SubjectUserRightsValidator.assertCanEdit(user, subject);
         return get(user, subjectService.save(user, subject, newSubject));
     }
-    
+
     @PutMapping("saveAndConfirm/{id:\\d+}")
-    public SubjectDto saveAndConfirm(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Subject subject, @Valid @RequestBody SubjectForm newSubject) {
-        UserUtil.assertSameSchool(user, subject.getSchool());
+    public SubjectDto saveAndConfirm(HoisUserDetails user, @WithVersionedEntity(versionRequestBody = true) Subject subject, @Valid @RequestBody SubjectForm newSubject) {
+        SubjectUserRightsValidator.assertCanEdit(user, subject);
+        SubjectUserRightsValidator.assertCanSetActive(user, subject);
         return get(user, subjectService.saveAndConfirm(user, subject, newSubject));
     }
-    
+
     @PutMapping("saveAndUnconfirm/{id:\\d+}")
-    public SubjectDto saveAndUnconfirm(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) Subject subject, @Valid @RequestBody SubjectForm newSubject) {
-        UserUtil.assertSameSchool(user, subject.getSchool());
+    public SubjectDto saveAndUnconfirm(HoisUserDetails user, @WithVersionedEntity(versionRequestBody = true) Subject subject, @Valid @RequestBody SubjectForm newSubject) {
+        SubjectUserRightsValidator.assertCanEdit(user, subject);
+        SubjectUserRightsValidator.assertCanSetPassive(user, subject);
         return get(user, subjectService.saveAndUnconfirm(user, subject, newSubject));
     }
 
     @GetMapping("/{id:\\d+}")
-    public SubjectDto get(HoisUserDetails user, @WithEntity("id") Subject subject) {
-        UserUtil.assertSameSchool(user, subject.getSchool());
-        return SubjectDto.of(subject, curriculumVersionRepository.findAllDistinctByModules_Subjects_Subject_id(subject.getId()));
+    public SubjectDto get(HoisUserDetails user, @WithEntity Subject subject) {
+        SubjectUserRightsValidator.assertCanView(user, subject);
+        return subjectService.get(user, subject);
     }
 
     @GetMapping
     public Page<SubjectSearchDto> search(@Valid SubjectSearchCommand subjectSearchCommand, HoisUserDetails user, Pageable pageable) {
-        return subjectService.search(user.getSchoolId(), subjectSearchCommand, pageable);
+        SubjectUserRightsValidator.assertCanSearch(user);
+        return subjectService.search(user, subjectSearchCommand, pageable);
     }
 
     @GetMapping("/initSearchFormData")
@@ -93,11 +111,11 @@ public class SubjectController {
     }
 
     @DeleteMapping("/{id:\\d+}")
-    public void delete(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestParam = "version") Subject subject, @SuppressWarnings("unused") @RequestParam("version") Long version) {
-        UserUtil.assertSameSchool(user, subject.getSchool());
+    public void delete(HoisUserDetails user, @WithVersionedEntity(versionRequestParam = "version") Subject subject, @SuppressWarnings("unused") @RequestParam("version") Long version) {
+        SubjectUserRightsValidator.assertCanDelete(user, subject);
         subjectService.delete(user, subject);
     }
-    
+
     @GetMapping("/unique/code")
     public boolean isCodeUnique(HoisUserDetails user, UniqueCommand command) {
         return subjectService.isCodeUnique(user.getSchoolId(), command);
